Added separate components and improved styles
Added ProductDescription as a component
Added ProductSection as a component
Improved SCSS
