+++ src/components/Accessoires/Accessoires.module.scss
@@ -111,7 +111,7 @@
     }
 
     &:hover {
-      color: white;
+      color: $color-white;
 
       &::after {
         color: rgba(255, 255, 255, 0.6);
@@ -124,7 +124,7 @@
   }
 
   span {
-    color: white;
+    color: $color-white;
     font-size: 14px;
     padding: 0.25rem 0;
   }
@@ -153,6 +153,9 @@
 .mainImageWrapper {
   width: 100%;
   max-width: 600px;
+  max-height: 600px;
+  border-radius: $border-radius;
+  overflow: hidden;
 }
 
 .infoSection {
@@ -168,7 +171,7 @@
 }
 
 .priceSection {
-  margin-bottom: 1.25rem;
+  margin-bottom: 1.5rem;
   .priceSale {
     font-size: $font-size-xl;
     font-weight: bold;
@@ -187,7 +190,8 @@
 }
 .quantitySection {
   display: flex;
-  align-items: center;
+  flex-direction: column;
+  align-items: flex-start;
   margin-bottom: 1.25rem;
   gap: 0.625rem;
 
@@ -241,7 +245,7 @@
   font-size: $font-size-lg;
   font-weight: 400;
 
-  margin: 0.625rem 0 0;
+  margin: 0 0 1.25rem;
   padding: 0 1.5rem;
   background: linear-gradient(135deg, $color-primary, $color-secondary);
   border: none;
@@ -290,7 +294,7 @@
 }
 
 .descriptionSection {
-  margin-top: 1.875rem;
+  //margin-top: 1.875rem;
 }
 
 .sectionTitle {
@@ -375,15 +379,23 @@
 }
 
 .toggleButton {
-  color: $color-white;
+  color: rgba(255, 255, 255, 0.6);
+  margin: 0 0 0.625rem;
+  transition: color 0.3s ease;
   cursor: pointer;
   font-size: $font-size-base;
-  margin-top: 0.625rem;
   background: none;
   border: none;
   padding: 0;
-  text-decoration: underline;
-  transition: none;
+  //text-decoration: underline;
+
+  &:hover {
+    color: white;
+
+    &::after {
+      color: rgba(255, 255, 255, 0.6);
+    }
+  }
 }
 
 .resetButton {
